wip: adiciona idSimulacao na mensagem de retorno

+++ src/main/java/br/com/hackathon/dto/simulacao/RespostaSimulacaoDto.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Data
 @Builder
 public class RespostaSimulacaoDto {
-    private Integer idSimulacao;
+    private Long idSimulacao;
     private Integer codigoProduto;
     private String descricaoProduto;
     private BigDecimal taxaJuros;
